requirements.txt
gradio
fastapi
uvicorn
requests
python-dotenv
